# Convert get_for_path to get_translatable