--- a/tests/views/unit.py
+++ b/tests/views/unit.py
@@ -9,7 +9,7 @@
 
 import pytest
 
-from pootle.core.exceptions import Http400
+from django.http import Http404
 
 from pootle_store.views import get_units
 
@@ -23,13 +23,13 @@ def test_get_units(rf, default):
 
     # `path` query parameter missing
     request = create_api_request(rf, user=default)
-    with pytest.raises(Http400):
+    with pytest.raises(Http404):
         view(request)
 
     # `path` query parameter present
     request = create_api_request(rf, url='/?path=foo', user=default)
-    response = view(request)
-    assert response.status_code == 200
+    with pytest.raises(Http404):
+        view(request)
 
 
 @pytest.mark.django_db