- replaces Unit.get_for_path with Unit.get_translatable
- avoids use of path or regex wherever possible