@phlax 2015-12-28
> Is it really necessary to alter the order of the parameters?

this method is useful for filtering out the template dirs for a user without any other filter - so it seems to make sense to switch the args and make the path optional.