@phlax 2015-12-28
> Shouldn't both commits be squashed?

maybe - but i expected some conversation about the second commit - so kept separate