@julen 2016-01-11
Unless I'm doing something wrong, testing this on InnoDB shows a massive performance degradation.

  | *master* | *pr/4317*
---- | ----- | -----
/ab/cd/export-view/?search=notebook&sfields=notes,locations |  12 queries run, total 0.125 seconds | 11 queries run, total 13.01 seconds
/ab/cd/export-view/?filter=incomplete | 16 queries run, total 0.191 seconds | 11 queries run, total 6.777 seconds
/ab/export-view/?filter=incomplete | 13 queries run, total 0.262 seconds | 11 queries run, total 6.861 seconds

Here */ab/cd/* is a TP with ~13000 words, and */ab/* contains ~16000 words. The degradation is noticeable in other views too, so this is not limited to the export view.

It might be the case that queries involving regexps do not use an index at all and they have to scan the entire table.