@phlax 2016-01-12
@julen i would expect a signifcant performance hit from moving to regexes.

I am not however seeing anything in the order that you are seeing above.

After repeated testing on our staging server against a lang with ~150000 live units the speed goes from ~.6 second to ~.8 second for the relevant regex query