@phlax 2016-01-12
...it could be that you are hitting resource limits in your env and this is forcing sorting etc on the disk, otherwise im not really understanding why there would be such a performance hit.