@phlax 2016-01-12
@julen would you be able to slowlog your mysql server to find out which queries are taking a long time