@julen 2016-01-12
This is in my local machine with far enough resources to handle this operation.

The slow log follows for accessing */ab/cd/export-view/?filter=incomplete*:
```sql
# Query_time: 2.678598  Lock_time: 0.003926 Rows_sent: 1  Rows_examined: 3200834
SELECT COUNT(*) AS `__count`
FROM `pootle_store_unit`
INNER JOIN `pootle_store_store` ON (`pootle_store_unit`.`store_id` = `pootle_store_store`.`id`)
WHERE (`pootle_store_unit`.`state` > -100
       AND NOT (`pootle_store_store`.`pootle_path` LIKE BINARY '/templates/%')
       AND `pootle_store_store`.`pootle_path` REGEXP BINARY '^/ab/cd/'
       AND (`pootle_store_unit`.`state` = 0
            OR `pootle_store_unit`.`state` = 50));

# Query_time: 4.785047  Lock_time: 0.000177 Rows_sent: 1732  Rows_examined: 3200834
SELECT `pootle_store_unit`.`id`,
       `pootle_store_unit`.`store_id`,
       `pootle_store_unit`.`index`,
       `pootle_store_unit`.`unitid`,
       `pootle_store_unit`.`unitid_hash`,
       `pootle_store_unit`.`source_f`,
       `pootle_store_unit`.`source_hash`,
       `pootle_store_unit`.`source_wordcount`,
       `pootle_store_unit`.`source_length`,
       `pootle_store_unit`.`target_f`,
       `pootle_store_unit`.`target_wordcount`,
       `pootle_store_unit`.`target_length`,
       `pootle_store_unit`.`developer_comment`,
       `pootle_store_unit`.`translator_comment`,
       `pootle_store_unit`.`locations`,
       `pootle_store_unit`.`context`,
       `pootle_store_unit`.`state`,
       `pootle_store_unit`.`revision`,
       `pootle_store_unit`.`creation_time`,
       `pootle_store_unit`.`mtime`,
       `pootle_store_unit`.`submitted_by_id`,
       `pootle_store_unit`.`submitted_on`,
       `pootle_store_unit`.`commented_by_id`,
       `pootle_store_unit`.`commented_on`,
       `pootle_store_unit`.`reviewed_by_id`,
       `pootle_store_unit`.`reviewed_on`,
       `pootle_store_unit`.`priority`,
       `pootle_store_store`.`id`,
       `pootle_store_store`.`file`,
       `pootle_store_store`.`parent_id`,
       `pootle_store_store`.`translation_project_id`,
       `pootle_store_store`.`pootle_path`,
       `pootle_store_store`.`name`,
       `pootle_store_store`.`file_mtime`,
       `pootle_store_store`.`state`,
       `pootle_store_store`.`creation_time`,
       `pootle_store_store`.`last_sync_revision`,
       `pootle_store_store`.`obsolete`
FROM `pootle_store_unit`
INNER JOIN `pootle_store_store` ON (`pootle_store_unit`.`store_id` = `pootle_store_store`.`id`)
WHERE (`pootle_store_unit`.`state` > -100
       AND NOT (`pootle_store_store`.`pootle_path` LIKE BINARY '/templates/%')
       AND `pootle_store_store`.`pootle_path` REGEXP BINARY '^/ab/cd/'
       AND (`pootle_store_unit`.`state` = 0
            OR `pootle_store_unit`.`state` = 50));
```

Note the `BINARY` keyword is being used because we have `utf8_bin` collation (cf. #3659). That doesn't affect performance though — and yeah, I checked temporarily switching to `utf8_general_ci` just in case.