@phlax 2016-01-12
@julen so im minded that we can probably get rid of the regexes altogether by not searching by path, we already have the parsed lang/proj/dir_path/filename - so there isnt really that much point to reassembling them in a regex/path.

however, im really keen to find out whats causing such significant differences in our 2 test environments.