@ta2-1 2016-01-14
@phlax, tested on translate and export views, it's working not worse. lgtm.