@julen 2016-01-15
Wonderful, this now works great. Well done @phlax! :+1: